Change the outputs of all BTOR verifiers to BTORVerifResult
